+++ bcos-protocol/bcos-protocol/protobuf/PBBlock.h
@@ -74,10 +74,10 @@ public:
     void encode(bytes& _encodeData) const override;
 
     // getNonces of the current block
-    Transaction::ConstPtr transaction(size_t _index) const override;
-    TransactionMetaData::ConstPtr transactionMetaData(size_t _index) const override;
+    Transaction::ConstPtr transaction(uint64_t _index) const override;
+    TransactionMetaData::ConstPtr transactionMetaData(uint64_t _index) const override;
 
-    TransactionReceipt::ConstPtr receipt(size_t _index) const override;
+    TransactionReceipt::ConstPtr receipt(uint64_t _index) const override;
 
     int32_t version() const override { return m_pbRawBlock->version(); }
 
@@ -110,7 +110,7 @@ public:
         clearTransactionsCache();
     }
     // Note: the caller must ensure the allocated transactions size
-    void setTransaction(size_t _index, Transaction::Ptr _transaction) override
+    void setTransaction(uint64_t _index, Transaction::Ptr _transaction) override
     {
         if (m_transactions->size() <= _index)
         {
@@ -132,7 +132,7 @@ public:
         clearReceiptsCache();
     }
     // Note: the caller must ensure the allocated receipts size
-    void setReceipt(size_t _index, TransactionReceipt::Ptr _receipt) override
+    void setReceipt(uint64_t _index, TransactionReceipt::Ptr _receipt) override
     {
         if (m_receipts->size() <= _index)
         {
@@ -153,11 +153,11 @@ public:
     }
 
     // get transactions size
-    size_t transactionsSize() const override { return m_transactions->size(); }
+    uint64_t transactionsSize() const override { return m_transactions->size(); }
     // get receipts size
-    size_t receiptsSize() const override { return m_receipts->size(); }
+    uint64_t receiptsSize() const override { return m_receipts->size(); }
 
-    size_t transactionsMetaDataSize() const override { return m_transactionMetaDataList->size(); }
+    uint64_t transactionsMetaDataSize() const override { return m_transactionMetaDataList->size(); }
     void setNonceList(NonceList const& _nonceList) override
     {
         *m_nonceList = _nonceList;
@@ -183,13 +183,13 @@ public:
         std::vector<bytes> transactionsList;
         if (transactionsSize() > 0)
         {
-            transactionsList = bcos::protocol::encodeToCalculateRoot(
-                transactionsSize(), [this](size_t _index) { return transaction(_index)->hash(); });
+            transactionsList = bcos::protocol::encodeToCalculateRoot(transactionsSize(),
+                [this](uint64_t _index) { return transaction(_index)->hash(); });
         }
         else if (transactionsMetaDataSize() > 0)
         {
             transactionsList = bcos::protocol::encodeToCalculateRoot(transactionsHashSize(),
-                [this](size_t _index) { return transactionMetaData(_index)->hash(); });
+                [this](uint64_t _index) { return transactionMetaData(_index)->hash(); });
         }
 
         txsRoot = bcos::protocol::calculateMerkleProofRoot(
@@ -207,7 +207,7 @@ public:
         }
 
         auto receiptsList = bcos::protocol::encodeToCalculateRoot(
-            receiptsSize(), [this](size_t _index) { return receipt(_index)->hash(); });
+            receiptsSize(), [this](uint64_t _index) { return receipt(_index)->hash(); });
         receiptsRoot =
             bcos::protocol::calculateMerkleProofRoot(m_receiptFactory->cryptoSuite(), receiptsList);
