+++ bcos-sealer/bcos-sealer/Sealer.h
@@ -42,10 +42,10 @@ public:
     void start() override;
     void stop() override;
 
-    void asyncNotifySealProposal(size_t _proposalStartIndex, size_t _proposalEndIndex,
-        size_t _maxTxsPerBlock, std::function<void(Error::Ptr)> _onRecvResponse) override;
+    void asyncNotifySealProposal(uint64_t _proposalStartIndex, uint64_t _proposalEndIndex,
+        uint64_t _maxTxsPerBlock, std::function<void(Error::Ptr)> _onRecvResponse) override;
     void asyncNoteUnSealedTxsSize(
-        size_t _unsealedTxsSize, std::function<void(Error::Ptr)> _onRecvResponse) override;
+        uint64_t _unsealedTxsSize, std::function<void(Error::Ptr)> _onRecvResponse) override;
 
     void asyncNoteLatestBlockNumber(int64_t _blockNumber) override;
     // interface for the consensus module to notify reset the sealing transactions
